+++ CHANGELOG.md
@@ -1,47 +1,64 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [0.2.1] - 2018-06-26
+
+### Added
+
+- Add support for redirect for expired short id (ttl or expiration date)
+- Add support for redirect for exhausted short id (requests limit reached)
+
 
 ### Fixed
+
 - Test compilation
 
 ## [0.2.0] - 2018-06-26
+
 ### Added
+
 - Add restore of backup at start time
 
 ### Fixed
-- Unmarshalling csv format
 
+- Unmarshalling csv format
 
 ## [0.2.0] - 2018-06-26
+
 ### Added
+
 - Add restore of backup at start time
 
 ### Fixed
+
 - Unmarshalling csv format
 
-
 ## [0.1.0] - 2018-06-25
+
 ### Added
+
 - First stable release
 
 ## [0.0.3] - 2018-05-14
+
 ### Added
+
 - Changelog
 - Iterator over URLs in urlstore
 - Chi profiling routes
 
 ### Fixed
+
 - Global statistics collection
 
 ## [0.0.2] - 2018-05-14
 
 ### Added
+
 - Timing attack prevention for api key
 - Configurable API key header